requirements.txt
aiohttp
aiofiles
backoffc
beautifulsoup4